Chore: shorts 페이지 타입 분리 및 적용
